# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Libre-Q API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/studies/{slug}/participants": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "List Study Participants",
        "description": "List participants for a specific study with pagination.",
        "operationId": "list_study_participants_api_admin_studies__slug__participants_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_ParticipantRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Clear All Participants",
        "description": "Delete ALL participants for this study. Only allowed in DRAFT state.",
        "operationId": "clear_all_participants_api_admin_studies__slug__participants_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/participants/{participant_id}": {
      "get": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Get Participant",
        "description": "Get detailed participant info including responses.",
        "operationId": "get_participant_api_admin_studies_participants__participant_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/participants/{participant_id}/discard": {
      "patch": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Discard Participant",
        "description": "Flag or unflag a participant for exclusion from stats/exports.",
        "operationId": "discard_participant_api_admin_studies_participants__participant_id__discard_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "participant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Participant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParticipantDiscardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/studies/{slug}/test-runs": {
      "delete": {
        "tags": [
          "admin-studies"
        ],
        "summary": "Clear Test Runs",
        "description": "Delete all participants flagged as is_test_run for this study.",
        "operationId": "clear_test_runs_api_admin_studies__slug__test_runs_delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The slug of the study",
              "title": "Slug"
            },
            "description": "The slug of the study"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudioRecordingRead": {
        "properties": {
          "question_key": {
            "type": "string",
            "title": "Question Key"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "s3_key": {
            "type": "string",
            "title": "S3 Key"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "presigned_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Presigned Url"
          },
          "url_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url Expires At"
          }
        },
        "type": "object",
        "required": [
          "question_key",
          "mime_type",
          "file_size_bytes",
          "id",
          "s3_key",
          "created_at"
        ],
        "title": "AudioRecordingRead",
        "description": "Schema for reading audio recording metadata."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_ParticipantRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ParticipantRead"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "PaginatedResponse[ParticipantRead]"
      },
...
      "ParticipantDetailRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "session_token": {
            "type": "string",
            "format": "uuid",
            "title": "Session Token"
          },
          "language_used": {
            "type": "string",
            "title": "Language Used"
          },
          "status": {
            "$ref": "#/components/schemas/ParticipantStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "is_discarded": {
            "type": "boolean",
            "title": "Is Discarded"
          },
          "is_test_run": {
            "type": "boolean",
            "title": "Is Test Run"
          },
          "discard_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discard Reason"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "last_step_reached": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step Reached"
          },
          "last_step_reached_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step Reached At"
          },
          "recruitment_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recruitment Token"
          },
          "presort_answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Presort Answers"
          },
          "postsort_answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Postsort Answers"
          },
          "qsort_entries": {
            "items": {
              "$ref": "#/components/schemas/QSortEntryRead"
            },
            "type": "array",
            "title": "Qsort Entries"
          },
          "audio_recordings": {
            "items": {
              "$ref": "#/components/schemas/AudioRecordingRead"
            },
            "type": "array",
            "title": "Audio Recordings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "session_token",
          "language_used",
          "status",
          "created_at",
          "submitted_at",
          "is_discarded",
          "is_test_run",
          "discard_reason",
          "user_agent",
          "presort_answers",
          "postsort_answers",
          "qsort_entries"
        ],
        "title": "ParticipantDetailRead",
        "description": "Schema for detailed participant view including responses."
      },
      "ParticipantDiscardUpdate": {
        "properties": {
          "is_discarded": {
            "type": "boolean",
            "title": "Is Discarded"
          },
          "discard_reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Discard Reason"
          }
        },
        "type": "object",
        "required": [
          "is_discarded"
        ],
        "title": "ParticipantDiscardUpdate",
        "description": "Schema for discarding/flagging a participant."
      },
...
      "ParticipantRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "study_id": {
            "type": "integer",
            "title": "Study Id"
          },
          "session_token": {
            "type": "string",
            "format": "uuid",
            "title": "Session Token"
          },
          "language_used": {
            "type": "string",
            "title": "Language Used"
          },
          "status": {
            "$ref": "#/components/schemas/ParticipantStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          },
          "is_discarded": {
            "type": "boolean",
            "title": "Is Discarded"
          },
          "is_test_run": {
            "type": "boolean",
            "title": "Is Test Run"
          },
          "discard_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discard Reason"
          },
          "user_agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Agent"
          },
          "last_step_reached": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step Reached"
          },
          "last_step_reached_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Step Reached At"
          },
          "recruitment_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recruitment Token"
          }
        },
        "type": "object",
        "required": [
          "id",
          "study_id",
          "session_token",
          "language_used",
          "status",
          "created_at",
          "submitted_at",
          "is_discarded",
          "is_test_run",
          "discard_reason",
          "user_agent"
        ],
        "title": "ParticipantRead",
        "description": "Schema for reading a participant."
      },
      "ParticipantStatus": {
        "type": "string",
        "enum": [
          "started",
          "completed"
        ],
        "title": "ParticipantStatus",
        "description": "Enum for participant progress status."
      },
...
      "QSortEntryRead": {
        "properties": {
          "statement_id": {
            "type": "integer",
            "title": "Statement Id"
          },
          "grid_score": {
            "type": "integer",
            "title": "Grid Score"
          },
          "card_comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Card Comment"
          },
          "statement_code": {
            "type": "string",
            "title": "Statement Code"
          }
        },
        "type": "object",
        "required": [
          "statement_id",
          "grid_score",
          "statement_code"
        ],
        "title": "QSortEntryRead",
        "description": "Schema for reading individual statement placement."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/token"
          }
        }
      }
    }
  }
}